requirements.txt
torch==2.6.0
trl==0.14.0
llama-stack==0.1.1
datasets==3.2.0
python-dotenv==1.0.1
tiktoken==0.8.0
